Refactor pcapng: move Custom to its own file
